pyproject.toml
[project]
name = "rh-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "inquirerpy>=0.3.4",
    "keyboard>=0.13.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "pyproject-toml>=0.1.0",
    "pyscreeze>=1.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "selenium>=4.28.1",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.394",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
]
